+++ rs/nns/governance/src/lib.rs
@@ -156,7 +156,7 @@ mod garbage_collection;
 /// subnetworks that participate in the Internet Computer (IC).
 pub mod governance;
 pub mod governance_proto_builder;
-pub mod heap_governance_data;
+mod heap_governance_data;
 mod known_neuron_index;
 mod migrations;
 mod neuron;
@@ -165,6 +165,7 @@ mod neuron_store;
 pub mod neurons_fund;
 mod node_provider_rewards;
 pub mod pb;
+mod proposal_data_neurons_fund_migration;
 pub mod proposals;
 mod reward;
 pub mod storage;
